// src/app/models/Persona.ts
export class Persona {
    id?: number;
    activo?: boolean;
    cedula?: string;
    celular?: string;
    correo?: string;
    nombres?: string;
    apellidos?:string;
    direccion?:string;
    calificacion?:string;
    tipo?:number;
    password?: string;
    fenixId?: number;
    authStatus?: string
}